+++ nix/default.nix
@@ -13,7 +13,7 @@ import sources.nixpkgs {
   overlays = [
     (_: pkgs: dapptools) # use released version to hit the binary cache
     (final: pkgs: rec {
-      go = go_1_20;
+      go = pkgs.go_1_20;
       go-ethereum = pkgs.callPackage ./go-ethereum.nix {
         inherit (pkgs.darwin) libobjc;
         inherit (pkgs.darwin.apple_sdk.frameworks) IOKit;