+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## UNRELEASED
+
+- [#986](https://github.com/crypto-org-chain/cronos/pull/986) Use go 1.20.
+
 *April 13, 2023*
 
 ## v1.0.7