+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASED
 
+- [#986](https://github.com/crypto-org-chain/cronos/pull/986) Use go 1.20.
 - [#984](https://github.com/crypto-org-chain/cronos/pull/984) experimental integration of memiavl.
 
 *April 13, 2023*