+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - [#934](https://github.com/crypto-org-chain/cronos/pull/934) Add pebbledb backend.
 - [#950](https://github.com/crypto-org-chain/cronos/pull/950) Implement memiavl and integrate with state machine.
 - [#985](https://github.com/crypto-org-chain/cronos/pull/985) Fix versiondb verify command on older versions
+- [#]() Bump grocksdb to v1.7.16 and rocksdb to v7.10.2
 
 *Feb 09, 2022*